Shell of app for SESG Mobile App
Current task for the Software Engineering Study Group is to develop any
mobile app of our choosing, any technology, any device(s). The idea for
this app is a Twitter style RESTful API to display a list of millions of
users, consumable by mobile devices. The mobile app will provide the
ability to scroll through the list of millions of users.